🎨 improve: refactor CLI structure by splitting command logic into separate modules and adding utility functions
