+++ src/cli/index.ts
@@ -1,0 +1,4 @@
+import { runMain as _runMain } from 'citty';
+import { mainCommand } from './command';
+
+export const runMain = () => _runMain(mainCommand);
